Moved FacilityCard files to FacilitiesManager folder.
