+++ WebApp/indoor-localization/src/components/facilities-manager/FacilityCard.tsx
